+++ client/CHANGELOG.md
@@ -1,3 +1,33 @@
+## client [1.38.1-rc.3](https://github.com/fhswf/appointme/compare/client@1.38.1-rc.2...client@1.38.1-rc.3) (2026-02-08)
+
+
+### Bug Fixes
+
+* improve highlighting of free slots ([a668551](https://github.com/fhswf/appointme/commit/a668551ee22e5a0e5ba36715183e2555153f0fbd))
+* improve highlighting of free slots ([1517144](https://github.com/fhswf/appointme/commit/1517144190c396e8bdc010ea8a8d4007bc3e5f6a))
+
+## client [1.38.1-rc.2](https://github.com/fhswf/appointme/compare/client@1.38.1-rc.1...client@1.38.1-rc.2) (2026-02-07)
+
+
+### Bug Fixes
+
+* improve highlighting of free slots ([c50e664](https://github.com/fhswf/appointme/commit/c50e66487da760d8e1b2b3f37cbdb18bfbea32cb))
+
+## client [1.38.1-rc.1](https://github.com/fhswf/appointme/compare/client@1.38.0...client@1.38.1-rc.1) (2026-02-07)
+
+
+### Bug Fixes
+
+* calculation of free slots in edge cases ([4c36f6f](https://github.com/fhswf/appointme/commit/4c36f6fa577ca7b306efc54d29e388f4a7362c74))
+
+
+
+
+
+### Dependencies
+
+* **common:** upgraded to 1.16.2-rc.1
+
 # client [1.38.0](https://github.com/fhswf/appointme/compare/client@1.37.2...client@1.38.0) (2026-01-29)